dashboard/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.2
streamlit==1.35.0
plotly==5.22.0
